requirements.txt
playwright==1.41.0
python-dateutil==2.8.2
pyyaml==6.0.1
apscheduler==3.10.4
